refactor(native): extract controller_mapping from keyboard (#2831)
Move the key->button mapping tables (gameboy/gba/snes_key_to_button_id),
handle_controller_key, and the NES P1/P2 button-routing helpers into
keyboard/controller_mapping.rs. Cross-module fns are pub(super); mod.rs
delegates via controller_mapping::. No behavior change.
